Refactor and enhance UdpTestServer utility.
Moved UdpTestServer to a new package and replaced volatile boolean with AtomicBoolean for better thread-safe state management. Added response update capabilities, improved timeout handling, and updated tests to reflect these changes. Enabled JVM arguments for Aeron unit tests in a following commit.
